+++ app_settings.csv
@@ -1,7 +1,7 @@
 key,value
-replicate_com_api_key,
-aws_access_key_id,
-aws_secret_access_key,
+replicate_com_api_key,061bfe6095279fe441cbcc33b310faa7b3a450eb
+aws_access_key_id,AKIAUVKRDGDXO3SJJK3I
+aws_secret_access_key,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
 previous_project,controlnet_test
-replicate_user_name,
-welcome_state,0
+replicate_user_name,piyushk52
+welcome_state,7
