+++ utils/file_upload/s3.py
@@ -1,0 +1,12 @@
+import boto3
+
+def upload_image(image_location):
+    app_settings = get_app_settings()
+
+    unique_file_name = str(uuid.uuid4()) + ".png"
+    s3_file = f"input_images/{unique_file_name}"
+    s3 = boto3.client('s3', aws_access_key_id=app_settings['aws_access_key_id'],
+                      aws_secret_access_key=app_settings['aws_secret_access_key'])
+    s3.upload_file(image_location, "banodoco", s3_file)
+    s3.put_object_acl(ACL='public-read', Bucket='banodoco', Key=s3_file)
+    return f"https://s3.amazonaws.com/banodoco/{s3_file}"
