+++ utils/ml_processor/ml_interface.py
@@ -1,0 +1,5 @@
+from abc import ABC
+
+
+class MachineLearningProcessor(ABC):
+    pass
